refactor: Replace index.ts with simpler version and move handlers to separate files
