+++ src/handlers/index.ts
@@ -4,3 +4,5 @@
  */
 
 export * from './projects.js';
+export * from './compliance-reports.js';
+export * from './quality-metrics.js';
